WIP input element
